Rename ru.spbau.alferov.javahw.hashable -> hashtable
